+++ src/pyport/api_client.py
@@ -36,794 +36,16 @@ Example:
         }
     )
     ```
+
+Note: The implementation has been refactored into the `client` package for better
+organization, but this module is maintained for backward compatibility.
 """
-import json
-import logging
-import os
-import random
-import threading
-import time
-from typing import Any, Callable, Dict, Optional, Set, Type, TypeVar, Union
 
+# Import from the refactored implementation
+from src.pyport.client.client import PortClient
+
+# Re-export for backward compatibility
+__all__ = ['PortClient']
+
+# Re-export modules used in tests
 import requests
-
-from src.pyport.retry import RetryConfig, RetryStrategy, with_retry
-
-from src.pyport.action_runs.action_runs_api_svc import ActionRuns
-from src.pyport.actions.actions_api_svc import Actions
-from src.pyport.apps.apps_api_svc import Apps
-from src.pyport.audit.audit_api_svc import Audit
-from src.pyport.checklist.checklist_api_svc import Checklist
-from src.pyport.constants import PORT_API_US_URL, PORT_API_URL, GENERIC_HEADERS
-from src.pyport.entities.entities_api_svc import Entities
-from src.pyport.error_handling import handle_error_response, handle_request_exception, with_error_handling
-from src.pyport.exceptions import (
-    PortApiError,
-    PortAuthenticationError, PortConfigurationError,
-    PortConnectionError,
-    PortPermissionError,
-    PortRateLimitError,
-    PortServerError,
-    PortTimeoutError
-)
-from src.pyport.integrations.integrations_api_svc import Integrations
-from src.pyport.logging import (
-    configure_logging,
-    log_request,
-    log_response,
-    log_error,
-    get_correlation_id,
-    logger as pyport_logger
-)
-from src.pyport.migrations.migrations_api_svc import Migrations
-from src.pyport.organization.organization_api_svc import Organizations
-from src.pyport.pages.pages_api_svc import Pages
-from src.pyport.blueprints.blueprint_api_svc import Blueprints
-from src.pyport.roles.roles_api_svc import Roles
-from src.pyport.scorecards.scorecards_api_svc import Scorecards
-from src.pyport.search.search_api_svc import Search
-from src.pyport.sidebars.sidebars_api_svc import Sidebars
-from src.pyport.teams.teams_api_svc import Teams
-from src.pyport.users.users_api_svc import Users
-
-# Remove imports for modules that don't exist yet
-# from src.pyport.data_sources.data_sources_api_svc import DataSources
-# from src.pyport.webhooks.webhooks_api_svc import Webhooks
-
-# Type variable for generic functions
-T = TypeVar('T')
-
-
-class PortClient:
-    """
-    Main client for interacting with the Port API.
-
-    This class provides a unified interface for all Port API operations,
-    handling authentication, request management, and error handling.
-    It exposes various service classes as properties for accessing
-    different parts of the API.
-
-    Attributes:
-        blueprints: Access to blueprint-related operations
-        entities: Access to entity-related operations
-        actions: Access to action-related operations
-        action_runs: Access to action run-related operations
-        pages: Access to page-related operations
-        integrations: Access to integration-related operations
-        organizations: Access to organization-related operations
-        teams: Access to team-related operations
-        users: Access to user-related operations
-        roles: Access to role-related operations
-        audit: Access to audit-related operations
-        migrations: Access to migration-related operations
-        search: Access to search-related operations
-        sidebars: Access to sidebar-related operations
-        checklist: Access to checklist-related operations
-        apps: Access to app-related operations
-        scorecards: Access to scorecard-related operations
-
-    Examples:
-        >>> # Create a client
-        >>> client = PortClient(
-        ...     client_id="your-client-id",
-        ...     client_secret="your-client-secret"
-        ... )
-        >>>
-        >>> # Get all blueprints
-        >>> blueprints = client.blueprints.get_blueprints()
-        >>>
-        >>> # Get a specific entity
-        >>> entity = client.entities.get_entity("service", "my-service")
-    """
-    def __init__(self, client_id: str, client_secret: str, us_region: bool = False,
-                 auto_refresh: bool = True, refresh_interval: int = 900,
-                 log_level: int = logging.INFO, log_format: Optional[str] = None,
-                 log_handler: Optional[logging.Handler] = None,
-                 # Retry configuration
-                 max_retries: int = 3,
-                 retry_delay: float = 1.0,
-                 max_delay: float = 10.0,
-                 retry_strategy: Union[str, RetryStrategy] = RetryStrategy.EXPONENTIAL,
-                 retry_jitter: bool = True,
-                 retry_status_codes: Optional[Set[int]] = None,
-                 retry_on: Optional[Union[Type[Exception], Set[Type[Exception]]]] = None,
-                 idempotent_methods: Optional[Set[str]] = None):
-        """
-        Initialize the PortClient.
-
-        Args:
-            client_id: API client ID obtained from Port.
-            client_secret: API client secret obtained from Port.
-            us_region: Whether to use the US region API URL (default: False).
-                Set to True if your Port instance is in the US region.
-            auto_refresh: If True, a background thread will refresh the token periodically (default: True).
-                Set to False if you want to manage token refresh manually.
-            refresh_interval: Token refresh interval in seconds (default: 900 sec = 15 minutes).
-                Tokens typically expire after 30 minutes, so refreshing every 15 minutes is recommended.
-            log_level: The logging level to use (default: logging.INFO).
-                Use logging.DEBUG for more detailed logs including request/response information.
-            log_format: The format string to use for log messages (default: None).
-                If None, a default format will be used: "%(asctime)s - %(name)s - %(levelname)s - %(message)s".
-            log_handler: A logging handler to use (default: None).
-                If None, a StreamHandler will be created that outputs to stderr.
-            max_retries: Maximum number of retry attempts for transient errors (default: 3).
-                Set to 0 to disable retries.
-            retry_delay: Initial delay between retries in seconds (default: 1.0).
-                This delay will be adjusted based on the retry strategy.
-            max_delay: Maximum delay between retries in seconds (default: 10.0).
-                No retry will wait longer than this, regardless of the strategy.
-            retry_strategy: Strategy for calculating retry delays (default: RetryStrategy.EXPONENTIAL).
-                Options: CONSTANT, LINEAR, EXPONENTIAL, FIBONACCI.
-            retry_jitter: Whether to add random jitter to retry delays (default: True).
-                Helps prevent thundering herd problems when multiple clients retry simultaneously.
-            retry_status_codes: HTTP status codes that should trigger retries (default: {429, 500, 502, 503, 504}).
-                Only applies to status codes returned by the API.
-            retry_on: Exception types or a function that returns True if the exception should be retried.
-                If None, uses default logic based on exception type and status code.
-            idempotent_methods: HTTP methods that are safe to retry (default: {"GET", "HEAD", "PUT", "DELETE", "OPTIONS"}).
-                Non-idempotent methods like POST are not retried by default to avoid duplicate operations.
-
-        Examples:
-            >>> # Basic client with default settings
-            >>> client = PortClient(
-            ...     client_id="your-client-id",
-            ...     client_secret="your-client-secret"
-            ... )
-            >>>
-            >>> # Client with custom retry settings
-            >>> client = PortClient(
-            ...     client_id="your-client-id",
-            ...     client_secret="your-client-secret",
-            ...     max_retries=5,
-            ...     retry_delay=0.5,
-            ...     retry_strategy=RetryStrategy.LINEAR
-            ... )
-            >>>
-            >>> # Client with custom logging
-            >>> import logging
-            >>> handler = logging.FileHandler("port_api.log")
-            >>> client = PortClient(
-            ...     client_id="your-client-id",
-            ...     client_secret="your-client-secret",
-            ...     log_level=logging.DEBUG,
-            ...     log_handler=handler
-            ... )
-        """
-        # Store authentication credentials
-        self.client_id = client_id
-        self.client_secret = client_secret
-
-        # Set up basic client properties
-        self._lock = threading.Lock()
-        self.api_url = PORT_API_US_URL if us_region else PORT_API_URL
-
-        # Configure components
-        self._setup_logging(log_level, log_format, log_handler)
-        self._setup_retry_config(max_retries, retry_delay, max_delay, retry_strategy, retry_jitter,
-                               retry_status_codes, retry_on, idempotent_methods)
-        self._setup_authentication()
-        self._setup_token_refresh(auto_refresh, refresh_interval)
-
-    def _setup_logging(self, log_level: int, log_format: Optional[str], log_handler: Optional[logging.Handler]) -> None:
-        """
-        Set up logging configuration.
-
-        Args:
-            log_level: The logging level to use.
-            log_format: The format string to use for log messages.
-            log_handler: A logging handler to use.
-        """
-        configure_logging(level=log_level, format_string=log_format, handler=log_handler)
-        self._logger = pyport_logger
-
-    def _setup_retry_config(self, max_retries: int, retry_delay: float, max_delay: float,
-                          retry_strategy: Union[str, RetryStrategy], retry_jitter: bool,
-                          retry_status_codes: Optional[Set[int]], retry_on: Optional[Union[Type[Exception], Set[Type[Exception]]]],
-                          idempotent_methods: Optional[Set[str]]) -> None:
-        """
-        Set up retry configuration.
-
-        Args:
-            max_retries: Maximum number of retry attempts.
-            retry_delay: Initial delay between retries in seconds.
-            max_delay: Maximum delay between retries in seconds.
-            retry_strategy: Strategy for calculating retry delays.
-            retry_jitter: Whether to add random jitter to retry delays.
-            retry_status_codes: HTTP status codes that should trigger retries.
-            retry_on: Exception types or function that determines if an exception should be retried.
-            idempotent_methods: HTTP methods that are safe to retry.
-        """
-        # Convert string strategy to enum if needed
-        if isinstance(retry_strategy, str):
-            retry_strategy = RetryStrategy(retry_strategy)
-
-        # Create retry configuration
-        self.retry_config = RetryConfig(
-            max_retries=max_retries,
-            retry_delay=retry_delay,
-            max_delay=max_delay,
-            strategy=retry_strategy,
-            jitter=retry_jitter,
-            retry_status_codes=retry_status_codes or {429, 500, 502, 503, 504},
-            retry_on=retry_on,
-            idempotent_methods=idempotent_methods or {"GET", "HEAD", "PUT", "DELETE", "OPTIONS"},
-            retry_hook=self._log_retry_attempt
-        )
-
-    def _setup_authentication(self) -> None:
-        """
-        Set up authentication by obtaining an access token and initializing the session and sub-clients.
-        """
-        # Obtain the initial token
-        self.token = self._get_access_token()
-
-        # Initialize the session and sub-clients
-        self._init_session()
-        self._init_sub_clients()
-
-    def _setup_token_refresh(self, auto_refresh: bool, refresh_interval: int) -> None:
-        """
-        Set up token refresh mechanism.
-
-        Args:
-            auto_refresh: Whether to automatically refresh the token.
-            refresh_interval: Token refresh interval in seconds.
-        """
-        self._auto_refresh = auto_refresh
-        self._refresh_interval = refresh_interval
-
-        # Start a background thread to auto-refresh the token if enabled
-        if self._auto_refresh:
-            self._start_token_refresh_thread()
-
-    def _init_session(self):
-        """Initializes the persistent session and default headers."""
-        self._session = requests.Session()
-        self._session.headers.update({
-            "Authorization": f"Bearer {self.token}",
-            "Content-Type": "application/json"
-        })
-
-    @property
-    def default_headers(self) -> dict:
-        """Return a copy of the default request headers."""
-        return dict(self._session.headers)
-
-    def _init_sub_clients(self):
-        """Initializes all API sub-clients."""
-        # Map of attribute names to their corresponding service classes
-        service_classes = {
-            'blueprints': Blueprints,
-            'entities': Entities,
-            'actions': Actions,
-            'pages': Pages,
-            'integrations': Integrations,
-            'action_runs': ActionRuns,
-            'organizations': Organizations,
-            'teams': Teams,
-            'users': Users,
-            'roles': Roles,
-            'audit': Audit,
-            'migrations': Migrations,
-            'search': Search,
-            'sidebars': Sidebars,
-            'checklist': Checklist,
-            'apps': Apps,
-            'scorecards': Scorecards,
-        }
-
-        # Initialize each service class
-        for attr_name, service_class in service_classes.items():
-            setattr(self, attr_name, service_class(self))
-
-        # Future services (commented out for now)
-        # 'webhooks': Webhooks,
-        # 'data_sources': DataSources,
-
-    def _start_token_refresh_thread(self):
-        refresh_thread = threading.Thread(target=self._token_refresh_loop, daemon=True)
-        refresh_thread.start()
-        self._logger.info("Token refresh thread started.")
-
-    def _handle_token_refresh_error(self, error: Exception) -> None:
-        """
-        Handle errors that occur during token refresh.
-
-        :param error: The exception that occurred.
-        """
-        if isinstance(error, PortRateLimitError):
-            # For rate limiting, respect the retry-after header
-            retry_after = error.retry_after or self._refresh_interval
-            self._logger.warning(
-                f"Rate limit exceeded during token refresh. "
-                f"Will retry after {retry_after} seconds. Error: {str(error)}"
-            )
-            # Sleep for the retry-after period minus the time already slept
-            time.sleep(max(1, retry_after - 1))
-        elif isinstance(error, (PortAuthenticationError, PortPermissionError)):
-            # Authentication errors are critical and should be logged as errors
-            self._logger.error(
-                f"Authentication error during token refresh. "
-                f"Check your client credentials. Error: {str(error)}"
-            )
-            # Sleep for a longer period before retrying
-            time.sleep(min(300, self._refresh_interval * 2))  # Max 5 minutes
-        elif isinstance(error, (PortServerError, PortTimeoutError, PortConnectionError)):
-            # Transient errors, log as warnings and retry normally
-            self._logger.warning(f"Transient error during token refresh: {str(error)}")
-        elif isinstance(error, PortApiError):
-            # Other API errors
-            self._logger.error(f"API error during token refresh: {str(error)}")
-        else:
-            # Unexpected errors
-            self._logger.error(f"Unexpected error during token refresh: {str(error)}")
-
-    def _token_refresh_loop(self):
-        """
-        Background thread that periodically refreshes the access token.
-        """
-        while True:
-            # Wait for the refresh interval
-            time.sleep(self._refresh_interval)
-
-            # Attempt to refresh the token
-            self._refresh_token()
-
-    def _refresh_token(self):
-        """
-        Refresh the access token and update the session headers.
-        """
-        try:
-            self._logger.debug("Refreshing access token...")
-            new_token = self._get_access_token()
-
-            # Update the token and session headers
-            with self._lock:
-                self.token = new_token
-                self._session.headers.update({"Authorization": f"Bearer {self.token}"})
-
-            self._logger.info("Access token refreshed successfully.")
-        except Exception as e:
-            self._handle_token_refresh_error(e)
-
-    def _extract_token_from_response(self, response_data: Dict[str, Any], endpoint: str) -> str:
-        """Extract the access token from the response data."""
-        token = response_data.get('accessToken')
-        if not token:
-            self._logger.error("Access token not found in the response.")
-            raise PortApiError(
-                "Access token not present in the API response",
-                endpoint=endpoint,
-                method="POST",
-                response_body=response_data
-            )
-        return token
-
-    def _prepare_auth_request(self) -> tuple[str, Dict[str, str], str]:
-        """
-        Prepare the authentication request.
-
-        :return: A tuple of (endpoint, headers, payload).
-        """
-        # Get credentials if needed
-        if not self.client_id or not self.client_secret:
-            self.client_id, self.client_secret = self._get_local_env_cred()
-
-        # Prepare request data
-        headers = GENERIC_HEADERS
-        credentials = {'clientId': self.client_id, 'clientSecret': self.client_secret}
-        payload = json.dumps(credentials)
-        endpoint = "auth/access_token"
-
-        return endpoint, headers, payload
-
-    def _handle_auth_response(self, response: requests.Response, endpoint: str) -> str:
-        """
-        Handle the authentication response.
-
-        Args:
-            response: The response from the authentication request.
-            endpoint: The API endpoint.
-
-        Returns:
-            The access token.
-
-        Raises:
-            PortApiError: If the response is invalid or doesn't contain a token.
-        """
-        if response.status_code == 200:
-            try:
-                return self._extract_token_from_response(response.json(), endpoint)
-            except json.JSONDecodeError as e:
-                error = PortApiError(
-                    "Invalid JSON response from authentication endpoint",
-                    endpoint=endpoint,
-                    method="POST"
-                )
-                log_error(error)
-                raise error from e
-
-        # Handle error response
-        error = handle_error_response(response, endpoint, "POST")
-        log_error(error)
-        raise error
-
-    def _get_access_token(self) -> str:
-        """
-        Get an access token from the API.
-
-        Returns:
-            The access token.
-
-        Raises:
-            PortAuthenticationError: If authentication fails.
-            PortApiError: If another API error occurs.
-            PortConfigurationError: If client credentials are missing.
-        """
-        # Generate a correlation ID for this request
-        correlation_id = get_correlation_id()
-
-        try:
-            # Prepare and send the authentication request
-            return self._send_auth_request(correlation_id)
-        except (PortApiError, PortAuthenticationError, PortConfigurationError):
-            # Re-raise these exceptions as they're already properly formatted
-            raise
-        except Exception as e:
-            # Catch any other exceptions and convert to PortApiError
-            return self._handle_unexpected_auth_error(e, correlation_id)
-
-    def _send_auth_request(self, correlation_id: str) -> str:
-        """
-        Prepare and send the authentication request.
-
-        Args:
-            correlation_id: A correlation ID for tracking the request.
-
-        Returns:
-            The access token.
-
-        Raises:
-            PortApiError: If the request fails.
-        """
-        # Prepare request
-        endpoint, headers, payload = self._prepare_auth_request()
-        url = f'{self.api_url}/v1/{endpoint}'
-
-        self._logger.debug("Sending authentication request to obtain access token...")
-
-        # Log the request (masking sensitive data)
-        log_request("POST", url, headers=headers, json_data=json.loads(payload),
-                    correlation_id=correlation_id)
-
-        try:
-            # Make the request
-            response = requests.post(url, headers=headers, data=payload, timeout=10)
-
-            # Log the response
-            log_response(response, correlation_id)
-
-            return self._handle_auth_response(response, endpoint)
-        except requests.RequestException as e:
-            # Convert requests exceptions to Port exceptions
-            error = handle_request_exception(e, endpoint, "POST")
-            log_error(error, correlation_id)
-            raise error
-
-    def _handle_unexpected_auth_error(self, e: Exception, correlation_id: str) -> str:
-        """
-        Handle unexpected errors during authentication.
-
-        Args:
-            e: The exception that occurred.
-            correlation_id: A correlation ID for tracking the request.
-
-        Raises:
-            PortApiError: A formatted API error with context about the original exception.
-        """
-        self._logger.error(f"An unexpected error occurred while obtaining access token: {str(e)}")
-        error = PortApiError(
-            f"Unexpected error during authentication: {str(e)}",
-            endpoint="auth/access_token",
-            method="POST"
-        )
-        log_error(error, correlation_id)
-        raise error from e
-
-    def _get_local_env_cred(self):
-        """
-        Get client credentials from environment variables.
-
-        Returns:
-            A tuple of (client_id, client_secret).
-
-        Raises:
-            PortConfigurationError: If credentials are missing.
-        """
-        # Get credentials from environment variables
-        client_id = os.getenv("PORT_CLIENT_ID")
-        client_secret = os.getenv("PORT_CLIENT_SECRET")
-
-        # Validate credentials
-        self._validate_credentials(client_id, client_secret, "environment variables")
-
-        return client_id, client_secret
-
-    def _validate_credentials(self, client_id: Optional[str], client_secret: Optional[str], source: str):
-        """
-        Validate that client credentials are present.
-
-        Args:
-            client_id: The client ID to validate.
-            client_secret: The client secret to validate.
-            source: The source of the credentials (for error messages).
-
-        Raises:
-            PortConfigurationError: If credentials are missing.
-        """
-        if not client_id or not client_secret:
-            self._logger.error(f"Missing credentials from {source}.")
-            raise PortConfigurationError(f"Client ID or client secret not found in {source}")
-
-    def _handle_response(self, response: requests.Response, endpoint: str, method: str, correlation_id: str) -> requests.Response:
-        """Handle the response, returning it if successful or raising an appropriate exception."""
-        # Log the response
-        log_response(response, correlation_id)
-
-        if 200 <= response.status_code < 300:
-            return response
-
-        # Handle error response
-        error = handle_error_response(response, endpoint, method)
-        log_error(error, correlation_id)
-        raise error
-
-    def _log_retry_attempt(self, error: Exception, attempt: int, delay: float) -> None:
-        """Log information about a retry attempt."""
-        self._logger.warning(
-            f"Retry attempt {attempt + 1}/{self.retry_config.max_retries} after {error.__class__.__name__}. "
-            f"Waiting {delay:.2f} seconds before retrying."
-        )
-
-        # Log additional details for API errors
-        if isinstance(error, PortApiError):
-            details = []
-            if error.status_code:
-                details.append(f"Status: {error.status_code}")
-            if error.endpoint:
-                details.append(f"Endpoint: {error.endpoint}")
-            if error.method:
-                details.append(f"Method: {error.method}")
-
-            if details:
-                self._logger.debug(f"Error details: {', '.join(details)}")
-
-    def _make_single_request(self, method: str, url: str, endpoint: str, correlation_id: str = None, **kwargs) -> requests.Response:
-        """
-        Make a single HTTP request and handle the response.
-
-        Args:
-            method: HTTP method (e.g., 'GET', 'POST').
-            url: The full URL to request.
-            endpoint: The API endpoint (for error reporting).
-            correlation_id: A correlation ID for tracking the request.
-            **kwargs: Additional parameters passed to requests.request.
-
-        Returns:
-            The response if successful.
-
-        Raises:
-            PortApiError: If the request fails.
-        """
-        # Generate or use the provided correlation ID
-        if correlation_id is None:
-            correlation_id = get_correlation_id()
-
-        # Log the request
-        log_request(method, url, headers=kwargs.get('headers'), params=kwargs.get('params'),
-                    data=kwargs.get('data'), json_data=kwargs.get('json'), correlation_id=correlation_id)
-        try:
-            # Make the request
-            response = self._session.request(method, url, **kwargs)
-
-            # Handle the response
-            return self._handle_response(response, endpoint, method, correlation_id)
-        except requests.RequestException as e:
-            # Convert requests exceptions to Port exceptions
-            error = handle_request_exception(e, endpoint, method)
-            log_error(error, correlation_id)
-            raise error
-
-    def make_request(self, method: str, endpoint: str, retries: int = None, retry_delay: float = None,
-                     correlation_id: str = None, **kwargs) -> requests.Response:
-        """
-        Make an HTTP request to the API with error handling and retry logic.
-
-        This is the main method for making API requests. It handles authentication,
-        request preparation, error handling, and retry logic. All API service classes
-        use this method to communicate with the Port API.
-
-        Args:
-            method: HTTP method (e.g., 'GET', 'POST', 'PUT', 'DELETE').
-                Different methods have different semantics:
-                - GET: Retrieve resources (idempotent)
-                - POST: Create resources (not idempotent)
-                - PUT: Replace resources (idempotent)
-                - PATCH: Update resources (not idempotent)
-                - DELETE: Remove resources (idempotent)
-            endpoint: API endpoint appended to the base URL.
-                For example, "blueprints" or "blueprints/{blueprint_id}".
-                The method automatically adds the API version prefix (e.g., "/v1/").
-            retries: Number of retry attempts for transient errors.
-                If None, uses the client's default (self.retry_config.max_retries).
-                Set to 0 to disable retries for this specific request.
-            retry_delay: Initial delay between retries in seconds (e.g., 1.0 = 1 second).
-                If None, uses the client's default (self.retry_config.retry_delay).
-                This delay will be adjusted based on the retry strategy.
-            correlation_id: A correlation ID for tracking the request.
-                If None, a new ID will be generated.
-                This ID is included in logs and can be used to trace a request through the system.
-            **kwargs: Additional parameters passed to requests.request.
-                Common parameters include:
-                - params: Dict of URL parameters
-                - json: Dict to be serialized as JSON in the request body
-                - data: Dict or string to be sent in the request body
-                - headers: Dict of HTTP headers to add/override
-                - timeout: Request timeout in seconds
-
-        Returns:
-            A requests.Response object containing the API response.
-            Use response.json() to get the parsed JSON content.
-
-        Raises:
-            PortAuthenticationError: If authentication fails.
-            PortResourceNotFoundError: If the requested resource doesn't exist.
-            PortValidationError: If the request data is invalid.
-            PortPermissionError: If the client doesn't have permission.
-            PortRateLimitError: If the API rate limit is exceeded.
-            PortServerError: If the server returns a 5xx error.
-            PortTimeoutError: If the request times out.
-            PortConnectionError: If there's a network connection error.
-            PortApiError: Base class for all Port API errors.
-
-        Examples:
-            >>> # Get all blueprints
-            >>> response = client.make_request('GET', 'blueprints')
-            >>> blueprints = response.json().get('blueprints', [])
-            >>>
-            >>> # Create a new blueprint
-            >>> blueprint_data = {
-            ...     "identifier": "microservice",
-            ...     "title": "Microservice"
-            ... }
-            >>> response = client.make_request('POST', 'blueprints', json=blueprint_data)
-            >>> new_blueprint = response.json().get('blueprint', {})
-            >>>
-            >>> # Get a specific blueprint with custom retry settings
-            >>> response = client.make_request(
-            ...     'GET',
-            ...     'blueprints/microservice',
-            ...     retries=5,
-            ...     retry_delay=0.5
-            ... )
-            >>> blueprint = response.json().get('blueprint', {})
-        """
-        # Generate or use the provided correlation ID
-        if correlation_id is None:
-            correlation_id = get_correlation_id()
-
-        # Build the full URL for the request
-        url = self._build_request_url(endpoint)
-
-        # Create a retry configuration for this request
-        local_config = self._create_request_retry_config(retries, retry_delay)
-
-        # Create a function with retry handling and execute it
-        return self._execute_request_with_retry(method, url, endpoint, correlation_id, local_config, **kwargs)
-
-    def _build_request_url(self, endpoint: str) -> str:
-        """
-        Build the full URL for a request based on the endpoint.
-
-        Args:
-            endpoint: The API endpoint to request.
-
-        Returns:
-            The full URL for the request.
-        """
-        # Check if we're running in a test environment
-        if 'test' in endpoint:
-            # For tests, don't add the /v1/ prefix
-            return f"{self.api_url}/{endpoint}"
-        else:
-            # For real API calls, add the /v1/ prefix
-            return f"{self.api_url}/v1/{endpoint}"
-
-    def _create_request_retry_config(self, retries: Optional[int], retry_delay: Optional[float]) -> RetryConfig:
-        """
-        Create a retry configuration for a request.
-
-        Args:
-            retries: Number of retry attempts for transient errors.
-                If None, uses the client's default.
-            retry_delay: Initial delay between retries in seconds.
-                If None, uses the client's default.
-
-        Returns:
-            A RetryConfig object for the request.
-        """
-        if retries is not None or retry_delay is not None:
-            return RetryConfig(
-                max_retries=retries if retries is not None else self.retry_config.max_retries,
-                retry_delay=retry_delay if retry_delay is not None else self.retry_config.retry_delay,
-                strategy=self.retry_config.strategy,
-                jitter=self.retry_config.jitter,
-                retry_status_codes=self.retry_config.retry_status_codes,
-                retry_on=self.retry_config.retry_on,
-                idempotent_methods=self.retry_config.idempotent_methods,
-                retry_hook=self.retry_config.retry_hook
-            )
-        else:
-            return self.retry_config
-
-    def _execute_request_with_retry(self, method: str, url: str, endpoint: str,
-                                  correlation_id: str, retry_config: RetryConfig,
-                                  **kwargs) -> requests.Response:
-        """
-        Execute a request with retry handling.
-
-        Args:
-            method: HTTP method (e.g., 'GET', 'POST').
-            url: The full URL to request.
-            endpoint: The API endpoint (for error reporting).
-            correlation_id: A correlation ID for tracking the request.
-            retry_config: The retry configuration to use.
-            **kwargs: Additional parameters passed to requests.request.
-
-        Returns:
-            A requests.Response object containing the API response.
-        """
-        # Define a function that will make a single request
-        def _make_request_impl(method, url, endpoint, correlation_id, **request_kwargs):
-            return self._make_single_request(method, url, endpoint, correlation_id, **request_kwargs)
-
-        # Apply the retry decorator to the function
-        make_request_with_retry = with_retry(_make_request_impl, config=retry_config)
-
-        # Add method to kwargs for the retry condition check
-        kwargs['method'] = method
-
-        # Make the request with retry handling
-        return make_request_with_retry(method, url, endpoint, correlation_id, **kwargs)
-
-    def with_error_handling(self, func: Callable[..., T], *args, **kwargs) -> T:
-        """
-        Execute a function with standard error handling.
-
-        :param func: The function to execute.
-        :param args: Positional arguments to pass to the function.
-        :param kwargs: Keyword arguments to pass to the function.
-        :return: The result of the function.
-        :raises PortApiError: If an error occurs and is not handled.
-        """
-        return with_error_handling(func)(*args, **kwargs)
