+++ tests/test_api_client.py
@@ -16,10 +16,9 @@ class TestPortClient(unittest.TestCase):
         os.environ.pop('PORT_CLIENT_ID', None)
         os.environ.pop('PORT_CLIENT_SECRET', None)
 
-    @patch('src.pyport.api_client.PortClient._get_access_token', return_value='dummy_token')
-    def test_initialization(self, mock_get_token):
+    def test_initialization(self):
         """Test that PortClient initializes sub-clients and session headers correctly."""
-        client = PortClient(client_secret="dummy_secret", client_id="dummy_id", us_region=True)
+        client = PortClient(client_secret="dummy_secret", client_id="dummy_id", us_region=True, skip_auth=True)
         self.assertEqual(client.token, 'dummy_token')
         self.assertEqual(client._session.headers.get("Authorization"), "Bearer dummy_token")
         self.assertIsNotNone(client.blueprints)
@@ -52,14 +51,13 @@ class TestPortClient(unittest.TestCase):
         os.environ['PORT_CLIENT_ID'] = 'dummy_id'
         os.environ['PORT_CLIENT_SECRET'] = 'dummy_secret'
 
-        client = PortClient(client_secret="dummy_secret", client_id="dummy_id", us_region=True)
+        client = PortClient(client_secret="dummy_secret", client_id="dummy_id", us_region=True, skip_auth=True)
         # Force calling _get_access_token by directly invoking it:
         token = client._get_access_token()
         self.assertEqual(token, expected_token)
 
-    @patch('src.pyport.api_client.PortClient._get_access_token', return_value='dummy_token')
-    @patch('src.pyport.api_client.requests.Session.request')
-    def test_make_request_success(self, mock_request, mock_get_token):
+    @patch('src.pyport.client.request.requests.Session.request')
+    def test_make_request_success(self, mock_request):
         """Test that make_request returns the expected JSON on success."""
         expected_json = {"key": "value"}
         dummy_response = MagicMock()
@@ -67,21 +65,20 @@ class TestPortClient(unittest.TestCase):
         dummy_response.json.return_value = expected_json
         mock_request.return_value = dummy_response
 
-        client = PortClient(client_secret="dummy_secret", client_id="dummy_id", us_region=True)
+        client = PortClient(client_secret="dummy_secret", client_id="dummy_id", us_region=True, skip_auth=True)
         response = client.make_request('GET', 'test-endpoint')
         mock_request.assert_called_once_with('GET', f"{client.api_url}/test-endpoint")
         self.assertEqual(response.json(), expected_json)
 
-    @patch('src.pyport.api_client.PortClient._get_access_token', return_value='dummy_token')
-    @patch('src.pyport.api_client.requests.Session.request')
-    def test_make_request_failure(self, mock_request, mock_get_token):
+    @patch('src.pyport.client.request.requests.Session.request')
+    def test_make_request_failure(self, mock_request):
         """Test that make_request raises an exception on HTTP error."""
         dummy_response = MagicMock()
         dummy_response.status_code = 404
         dummy_response.raise_for_status.side_effect = requests.HTTPError("Not Found")
         mock_request.return_value = dummy_response
 
-        client = PortClient(client_secret="dummy_secret", client_id="dummy_id", us_region=True)
+        client = PortClient(client_secret="dummy_secret", client_id="dummy_id", us_region=True, skip_auth=True)
         with self.assertRaises(PortResourceNotFoundError) as context:
             client.make_request('GET', 'nonexistent-endpoint')
         self.assertEqual(context.exception.status_code, 404)
